+++ Packs/ApiModules/Scripts/CSVFeedApiModule/CSVFeedApiModule.py
@@ -7,10 +7,12 @@ import csv
 import gzip
 import urllib3
 from dateutil.parser import parse
-from typing import Optional, Pattern, Dict, Any, Tuple, Union
+from typing import Optional, Pattern, Dict, Any, Tuple, Union, List
 
 # disable insecure warnings
 urllib3.disable_warnings()
+
+# Globals
 
 
 class Client(BaseClient):
@@ -18,7 +20,7 @@ class Client(BaseClient):
                  insecure: bool = False, credentials: dict = None, ignore_regex: str = None, encoding: str = 'latin-1',
                  delimiter: str = ',', doublequote: bool = True, escapechar: str = '',
                  quotechar: str = '"', skipinitialspace: bool = False, polling_timeout: int = 20, proxy: bool = False,
-                 **kwargs):
+                 feedTags: Optional[str] = None, **kwargs):
         """
         :param url: URL of the feed.
         :param feed_url_to_config: for each URL, a configuration of the feed that contains
@@ -60,6 +62,7 @@ class Client(BaseClient):
         :param polling_timeout: timeout of the polling request in seconds. Default: 20
         :param proxy: Sets whether use proxy when sending requests
         """
+        self.tags: List[str] = argToList(feedTags)
         if not credentials:
             credentials = {}
 
@@ -181,21 +184,25 @@ class Client(BaseClient):
         return response_content.decode(self.encoding).split('\n')
 
 
-def determine_indicator_type(indicator_type, default_indicator_type, value):
+def determine_indicator_type(indicator_type, default_indicator_type, auto_detect, value):
+    """
+    Detect the indicator type of the given value.
+    Args:
+        indicator_type: (str) Indicator type given in the config.
+        default_indicator_type: Indicator type which was inserted as a param of the integration by user.
+        auto_detect: (bool) True whether auto detection of the indicator type is wanted.
+        value: (str) The value which we'd like to get indicator type of.
+    Returns:
+        Str which stands for the indicator type after detection.
+    """
+    if auto_detect:
+        indicator_type = auto_detect_indicator_type(value)
     if not indicator_type:
         indicator_type = default_indicator_type
-    if indicator_type == FeedIndicatorType.Domain and '*' in value:
-        indicator_type = FeedIndicatorType.DomainGlob
     return indicator_type
 
 
 def module_test_command(client: Client, args):
-    if not client.feed_url_to_config:
-        indicator_type = args.get('indicator_type', demisto.params().get('indicator_type'))
-        if not FeedIndicatorType.is_valid_type(indicator_type):
-            supported_values = FeedIndicatorType.list_all_supported_indicators()
-            raise ValueError(f'Indicator type of {indicator_type} is not supported. Supported values are:'
-                             f' {supported_values}')
     client.build_iterator()
     return 'ok', {}, {}
 
@@ -240,7 +247,7 @@ def create_fields_mapping(raw_json: Dict[str, Any], mapping: Dict[str, Union[Tup
     return fields_mapping
 
 
-def fetch_indicators_command(client: Client, default_indicator_type: str, **kwargs):
+def fetch_indicators_command(client: Client, default_indicator_type: str, auto_detect: bool, **kwargs):
     iterator = client.build_iterator(**kwargs)
     indicators = []
     config = client.feed_url_to_config or {}
@@ -255,7 +262,8 @@ def fetch_indicators_command(client: Client, default_indicator_type: str, **kwar
                 if value:
                     raw_json['value'] = value
                     conf_indicator_type = config.get(url, {}).get('indicator_type')
-                    indicator_type = determine_indicator_type(conf_indicator_type, default_indicator_type, value)
+                    indicator_type = determine_indicator_type(conf_indicator_type, default_indicator_type, auto_detect,
+                                                              value)
                     raw_json['type'] = indicator_type
                     indicator = {
                         'value': value,
@@ -263,15 +271,21 @@ def fetch_indicators_command(client: Client, default_indicator_type: str, **kwar
                         'rawJSON': raw_json,
                         'fields': create_fields_mapping(raw_json, mapping) if mapping else {}
                     }
+                    indicator['fields']['tags'] = client.tags
                     indicators.append(indicator)
-
     return indicators
 
 
-def get_indicators_command(client, args):
+def get_indicators_command(client, args: dict, tags: Optional[List[str]] = None):
+    if tags is None:
+        tags = []
     itype = args.get('indicator_type', demisto.params().get('indicator_type'))
-    limit = int(args.get('limit'))
-    indicators_list = fetch_indicators_command(client, itype)
+    try:
+        limit = int(args.get('limit', 50))
+    except ValueError:
+        raise ValueError('The limit argument must be a number.')
+    auto_detect = demisto.params().get('auto_detect_type')
+    indicators_list = fetch_indicators_command(client, itype, auto_detect)
     entry_result = indicators_list[:limit]
     hr = tableToMarkdown('Indicators', entry_result, headers=['value', 'type', 'fields'])
     return hr, {}, indicators_list
@@ -294,7 +308,11 @@ def feed_main(feed_name, params=None, prefix=''):
     }
     try:
         if command == 'fetch-indicators':
-            indicators = fetch_indicators_command(client, params.get('indicator_type'))
+            indicators = fetch_indicators_command(
+                client,
+                params.get('indicator_type'),
+                params.get('auto_detect_type')
+            )
             # we submit the indicators in batches
             for b in batch(indicators, batch_size=2000):
                 demisto.createIndicators(b)  # type: ignore