+++ Packs/Base/Scripts/CommonServerPython/CommonServerPython.py
@@ -14,6 +14,7 @@ import socket
 import sys
 import time
 import traceback
+from random import randint
 import xml.etree.cElementTree as ET
 from collections import OrderedDict
 from datetime import datetime, timedelta
@@ -39,7 +40,7 @@ if IS_PY3:
     STRING_TYPES = (str, bytes)  # type: ignore
     STRING_OBJ_TYPES = (str,)
 else:
-    STRING_TYPES = (str, unicode)  # type: ignore
+    STRING_TYPES = (str, unicode)  # type: ignore # noqa: F821
     STRING_OBJ_TYPES = STRING_TYPES  # type: ignore
 # pylint: enable=undefined-variable
 
@@ -302,7 +303,7 @@ def auto_detect_indicator_type(indicator_value):
         return FeedIndicatorType.CVE
 
     try:
-        no_cache_extract = tldextract.TLDExtract(cache_file=False,suffix_list_urls=None)
+        no_cache_extract = tldextract.TLDExtract(cache_file=False, suffix_list_urls=None)
         if no_cache_extract(indicator_value).suffix:
             if '*' in indicator_value:
                 return FeedIndicatorType.DomainGlob
@@ -370,7 +371,7 @@ def handle_proxy(proxy_param_name='proxy', checkbox_default_value=False, handle_
         if insecure_param_name is None:
             param_names = ('insecure', 'unsecure')
         else:
-            param_names = (insecure_param_name, )
+            param_names = (insecure_param_name, )  # type: ignore[assignment]
         for p in param_names:
             if demisto.params().get(p, False):
                 for k in ('REQUESTS_CA_BUNDLE', 'CURL_CA_BUNDLE'):
@@ -837,7 +838,7 @@ def encode_string_results(text):
         return text
     try:
         return str(text)
-    except UnicodeEncodeError as exception:
+    except UnicodeEncodeError:
         return text.encode("utf8", "replace")
 
 
@@ -1193,7 +1194,7 @@ def argToList(arg, separator=','):
         if arg[0] == '[' and arg[-1] == ']':
             return json.loads(arg)
         return [s.strip() for s in arg.split(separator)]
-    return arg
+    return [arg]
 
 
 def argToBoolean(value):
@@ -1250,7 +1251,7 @@ def appendContext(key, data, dedup=False):
 
         elif isinstance(existing, dict):
             if isinstance(data, dict):
-                new_val = [existing, data]
+                new_val = [existing, data]  # type: ignore[assignment]
             else:
                 new_val = data + existing  # will raise a self explanatory TypeError
 
@@ -1259,10 +1260,10 @@ def appendContext(key, data, dedup=False):
                 existing.extend(data)
             else:
                 existing.append(data)
-            new_val = existing
+            new_val = existing  # type: ignore[assignment]
 
         else:
-            new_val = [existing, data]
+            new_val = [existing, data]  # type: ignore[assignment]
 
         if dedup and isinstance(new_val, list):
             new_val = list(set(new_val))
@@ -1339,7 +1340,7 @@ def tableToMarkdown(name, t, headers=None, headerTransform=None, removeNull=Fals
     if t and len(headers) > 0:
         newHeaders = []
         if headerTransform is None:  # noqa
-            headerTransform = lambda s: s  # noqa
+            def headerTransform(s): return s  # noqa
         for header in headers:
             newHeaders.append(headerTransform(header))
         mdResult += '|'
@@ -1391,7 +1392,7 @@ def createContextSingle(obj, id=None, keyTransform=None, removeNull=False):
     """
     res = {}  # type: dict
     if keyTransform is None:
-        keyTransform = lambda s: s  # noqa
+        def keyTransform(s): return s  # noqa
     keys = obj.keys()
     for key in keys:
         if removeNull and obj[key] in ('', None, [], {}):
@@ -1475,7 +1476,7 @@ def fileResult(filename, data, file_type=None):
         file_type = entryTypes['file']
     temp = demisto.uniqueFile()
     # pylint: disable=undefined-variable
-    if (IS_PY3 and isinstance(data, str)) or (not IS_PY3 and isinstance(data, unicode)):  # type: ignore
+    if (IS_PY3 and isinstance(data, str)) or (not IS_PY3 and isinstance(data, unicode)):  # type: ignore # noqa: F821
         data = data.encode('utf-8')
     # pylint: enable=undefined-variable
     with open(demisto.investigation()['id'] + '_' + temp, 'wb') as f:
@@ -2353,7 +2354,6 @@ class Common(object):
 
             return ret_value
 
-
     class Domain(Indicator):
         """ ignore docstring
         Domain indicator - https://xsoar.pan.dev/docs/context-standards#domain
@@ -2389,7 +2389,6 @@ class Common(object):
             self.admin_email = admin_email
             self.admin_phone = admin_phone
             self.admin_country = admin_country
-
 
             self.domain_status = domain_status
             self.name_servers = name_servers
@@ -2581,10 +2580,12 @@ class CommandResults:
     :return: None
     :rtype: ``None``
     """
-    def __init__(self, outputs_prefix, outputs_key_field, outputs, indicators=None, readable_output=None,
+    def __init__(self, outputs_prefix=None, outputs_key_field=None, outputs=None, indicators=None, readable_output=None,
                  raw_response=None):
-
         # type: (str, str, object, list, str, object) -> None
+        if raw_response is None:
+            raw_response = outputs
+
         self.indicators = indicators
 
         self.outputs_prefix = outputs_prefix
@@ -2596,8 +2597,11 @@ class CommandResults:
 
     def to_context(self):
         outputs = {}  # type: dict
-        human_readable = None
-        raw_response = None
+        if self.readable_output:
+            human_readable = self.readable_output
+        else:
+            human_readable = None  # type: ignore[assignment]
+        raw_response = None  # type: ignore[assignment]
 
         if self.indicators:
             for indicator in self.indicators:
@@ -2612,16 +2616,10 @@ class CommandResults:
         if self.raw_response:
             raw_response = self.raw_response
 
-        if self.outputs:
+        if self.outputs is not None:
             if not self.readable_output:
                 # if markdown is not provided then create table by default
                 human_readable = tableToMarkdown('Results', self.outputs)
-            else:
-                human_readable = self.readable_output
-
-            if not self.raw_response:
-                raw_response = self.outputs
-
             if self.outputs_prefix and self.outputs_key_field:
                 # if both prefix and key field provided then create DT key
                 outputs_key = '{0}(val.{1} == obj.{1})'.format(self.outputs_prefix, self.outputs_key_field)
@@ -2630,8 +2628,7 @@ class CommandResults:
                 outputs_key = '{}'.format(self.outputs_prefix)
                 outputs[outputs_key] = self.outputs
             else:
-                outputs = self.outputs
-                human_readable = self.readable_output  # prefix and key field not provided, human readable should
+                outputs = self.outputs  # type: ignore[assignment]
 
         return_entry = {
             'Type': EntryType.NOTE,
@@ -2873,9 +2870,9 @@ regexFlags = re.M  # Multi line matching
 # else, use re.match({regex_format},str)
 
 ipv4Regex = r'\b((25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.){3}(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\b([^\/]|$)'
-ipv4cidrRegex = r'\b(?:(?:25[0-5]|2[0-4][0-9]|1[0-9][0-9]|[1-9]?[0-9])(?:\[\.\]|\.)){3}(?:25[0-5]|2[0-4][0-9]|1[0-9][0-9]|[1-9]?[0-9])(\/([0-9]|[1-2][0-9]|3[0-2]))\b'
-ipv6Regex = r'\b(?:(?:[0-9a-fA-F]{1,4}:){7,7}[0-9a-fA-F]{1,4}|(?:[0-9a-fA-F]{1,4}:){1,7}:|(?:[0-9a-fA-F]{1,4}:){1,6}:[0-9a-fA-F]{1,4}|(?:[0-9a-fA-F]{1,4}:){1,5}(:[0-9a-fA-F]{1,4}){1,2}|(?:[0-9a-fA-F]{1,4}:){1,4}(:[0-9a-fA-F]{1,4}){1,3}|(?:[0-9a-fA-F]{1,4}:){1,3}(:[0-9a-fA-F]{1,4}){1,4}|(?:[0-9a-fA-F]{1,4}:){1,2}(:[0-9a-fA-F]{1,4}){1,5}|[0-9a-fA-F]{1,4}:((:[0-9a-fA-F]{1,4}){1,6})|:(?:(:[0-9a-fA-F]{1,4}){1,7}|:)|fe80:(:[0-9a-fA-F]{0,4}){0,4}%[0-9a-zA-Z]{1,}|::(ffff(:0{1,4}){0,1}:){0,1}((25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9])\.){3,3}(25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9])|([0-9a-fA-F]{1,4}:){1,4}:((25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9])\.){3,3}(25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9]))\b'
-ipv6cidrRegex = r'\b(([0-9a-fA-F]{1,4}:){7,7}[0-9a-fA-F]{1,4}|([0-9a-fA-F]{1,4}:){1,7}:|([0-9a-fA-F]{1,4}:){1,6}:[0-9a-fA-F]{1,4}|([0-9a-fA-F]{1,4}:){1,5}(:[0-9a-fA-F]{1,4}){1,2}|([0-9a-fA-F]{1,4}:){1,4}(:[0-9a-fA-F]{1,4}){1,3}|([0-9a-fA-F]{1,4}:){1,3}(:[0-9a-fA-F]{1,4}){1,4}|([0-9a-fA-F]{1,4}:){1,2}(:[0-9a-fA-F]{1,4}){1,5}|[0-9a-fA-F]{1,4}:((:[0-9a-fA-F]{1,4}){1,6})|:((:[0-9a-fA-F]{1,4}){1,7}|:)|fe80:(:[0-9a-fA-F]{0,4}){0,4}%[0-9a-zA-Z]{1,}|::(ffff(:0{1,4}){0,1}:){0,1}((25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9])\.){3,3}(25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9])|([0-9a-fA-F]{1,4}:){1,4}:((25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9])\.){3,3}(25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9]))(\/(12[0-8]|1[0-1][0-9]|[1-9][0-9]|[0-9]))\b'
+ipv4cidrRegex = r'\b(?:(?:25[0-5]|2[0-4][0-9]|1[0-9][0-9]|[1-9]?[0-9])(?:\[\.\]|\.)){3}(?:25[0-5]|2[0-4][0-9]|1[0-9][0-9]|[1-9]?[0-9])(\/([0-9]|[1-2][0-9]|3[0-2]))\b'  # noqa: E501
+ipv6Regex = r'\b(?:(?:[0-9a-fA-F]{1,4}:){7,7}[0-9a-fA-F]{1,4}|(?:[0-9a-fA-F]{1,4}:){1,7}:|(?:[0-9a-fA-F]{1,4}:){1,6}:[0-9a-fA-F]{1,4}|(?:[0-9a-fA-F]{1,4}:){1,5}(:[0-9a-fA-F]{1,4}){1,2}|(?:[0-9a-fA-F]{1,4}:){1,4}(:[0-9a-fA-F]{1,4}){1,3}|(?:[0-9a-fA-F]{1,4}:){1,3}(:[0-9a-fA-F]{1,4}){1,4}|(?:[0-9a-fA-F]{1,4}:){1,2}(:[0-9a-fA-F]{1,4}){1,5}|[0-9a-fA-F]{1,4}:((:[0-9a-fA-F]{1,4}){1,6})|:(?:(:[0-9a-fA-F]{1,4}){1,7}|:)|fe80:(:[0-9a-fA-F]{0,4}){0,4}%[0-9a-zA-Z]{1,}|::(ffff(:0{1,4}){0,1}:){0,1}((25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9])\.){3,3}(25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9])|([0-9a-fA-F]{1,4}:){1,4}:((25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9])\.){3,3}(25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9]))\b'  # noqa: E501
+ipv6cidrRegex = r'\b(([0-9a-fA-F]{1,4}:){7,7}[0-9a-fA-F]{1,4}|([0-9a-fA-F]{1,4}:){1,7}:|([0-9a-fA-F]{1,4}:){1,6}:[0-9a-fA-F]{1,4}|([0-9a-fA-F]{1,4}:){1,5}(:[0-9a-fA-F]{1,4}){1,2}|([0-9a-fA-F]{1,4}:){1,4}(:[0-9a-fA-F]{1,4}){1,3}|([0-9a-fA-F]{1,4}:){1,3}(:[0-9a-fA-F]{1,4}){1,4}|([0-9a-fA-F]{1,4}:){1,2}(:[0-9a-fA-F]{1,4}){1,5}|[0-9a-fA-F]{1,4}:((:[0-9a-fA-F]{1,4}){1,6})|:((:[0-9a-fA-F]{1,4}){1,7}|:)|fe80:(:[0-9a-fA-F]{0,4}){0,4}%[0-9a-zA-Z]{1,}|::(ffff(:0{1,4}){0,1}:){0,1}((25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9])\.){3,3}(25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9])|([0-9a-fA-F]{1,4}:){1,4}:((25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9])\.){3,3}(25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9]))(\/(12[0-8]|1[0-1][0-9]|[1-9][0-9]|[0-9]))\b'  # noqa: E501
 emailRegex = r'\b[^@]+@[^@]+\.[^@]+\b'
 hashRegex = r'\b[0-9a-fA-F]+\b'
 urlRegex = r'(?:(?:https?|ftp|hxxps?):\/\/|www\[?\.\]?|ftp\[?\.\]?)(?:[-\w\d]+\[?\.\]?)+[-\w\d]+(?::\d+)?' \
@@ -3055,7 +3052,7 @@ def parse_date_range(date_range, date_format=None, to_timestamp=False, timezone=
     return start_time, end_time
 
 
-def timestamp_to_datestring(timestamp, date_format="%Y-%m-%dT%H:%M:%S.000Z"):
+def timestamp_to_datestring(timestamp, date_format="%Y-%m-%dT%H:%M:%S.000Z", is_utc=False):
     """
       Parses timestamp (milliseconds) to a date string in the provided date format (by default: ISO 8601 format)
       Examples: (1541494441222, 1541495441000, etc.)
@@ -3066,9 +3063,15 @@ def timestamp_to_datestring(timestamp, date_format="%Y-%m-%dT%H:%M:%S.000Z"):
       :type date_format: ``str``
       :param date_format: The date format the timestamp should be parsed to. (optional)
 
+      :type is_utc: ``bool``
+      :param is_utc: Should the string representation of the timestamp use UTC time or the local machine time
+
       :return: The parsed timestamp in the date_format
       :rtype: ``str``
     """
+    use_utc_time = is_utc or date_format.endswith('Z')
+    if use_utc_time:
+        return datetime.utcfromtimestamp(int(timestamp) / 1000.0).strftime(date_format)
     return datetime.fromtimestamp(int(timestamp) / 1000.0).strftime(date_format)
 
 
@@ -3148,20 +3151,29 @@ def assign_params(keys_to_ignore=None, values_to_ignore=None, **kwargs):
     }
 
 
-def get_demisto_version():
-    """Returns the Demisto version and build number.
-
-    :return: Demisto version object if Demisto class has attribute demistoVersion, else raises AttributeError
-    :rtype: ``dict``
+class GetDemistoVersion:
     """
-    if getattr(get_demisto_version, '_version', None):
-        return get_demisto_version._version
-    if hasattr(demisto, 'demistoVersion'):
-        version = demisto.demistoVersion()
-        get_demisto_version._version = version
-        return version
-    else:
-        raise AttributeError('demistoVersion attribute not found.')
+    Callable class to replace get_demisto_version function
+    """
+
+    def __init__(self):
+        self._version = None
+
+    def __call__(self):
+        """Returns the Demisto version and build number.
+
+        :return: Demisto version object if Demisto class has attribute demistoVersion, else raises AttributeError
+        :rtype: ``dict``
+        """
+        if self._version is None:
+            if hasattr(demisto, 'demistoVersion'):
+                self._version = demisto.demistoVersion()
+            else:
+                raise AttributeError('demistoVersion attribute not found.')
+        return self._version
+
+
+get_demisto_version = GetDemistoVersion()
 
 
 def is_demisto_version_ge(version):
@@ -3575,7 +3587,7 @@ if 'requests' in sys.modules:
 
         def _http_request(self, method, url_suffix, full_url=None, headers=None, auth=None, json_data=None,
                           params=None, data=None, files=None, timeout=10, resp_type='json', ok_codes=None,
-                          return_empty_response = False, retries=0, status_list_to_retry=None,
+                          return_empty_response=False, retries=0, status_list_to_retry=None,
                           backoff_factor=5, raise_on_redirect=False, raise_on_status=False, **kwargs):
             """A wrapper for requests lib to send our requests and handle requests and responses better.
 
@@ -3739,11 +3751,10 @@ if 'requests' in sys.modules:
             except requests.exceptions.RetryError as exception:
                 try:
                     reason = 'Reason: {}'.format(exception.args[0].reason.args[0])
-                except:
+                except Exception:
                     reason = ''
                 err_msg = 'Max Retries Error- Request attempts with {} retries failed. \n{}'.format(retries, reason)
                 raise DemistoException(err_msg, exception)
-
 
         def _is_status_code_valid(self, response, ok_codes=None):
             """If the status code is OK, return 'True'.
@@ -3784,6 +3795,236 @@ def batch(iterable, batch_size=1):
         yield current_batch
         current_batch = not_batched[:batch_size]
         not_batched = not_batched[batch_size:]
+
+
+def dict_safe_get(dict_object, keys, default_return_value=None):
+    """Recursive safe get query, If keys found return value otherwise return None or default value.
+
+    :type dict_object: ``dict``
+    :param dict_object: dictionary to query.
+
+    :type keys: ``list``
+    :param keys: keys for recursive get.
+
+    :type default_return_value: ``object``
+    :param default_return_value: Value to return when no key available.
+
+    :rtype: ``object``
+    :return:: Value found.
+    """
+    for key in keys:
+        try:
+            dict_object = dict_object[key]
+        except (KeyError, TypeError):
+            return default_return_value
+
+    return dict_object
+
+
+CONTEXT_UPDATE_RETRY_TIMES = 3
+MIN_VERSION_FOR_VERSIONED_CONTEXT = '6.0.0'
+
+
+def merge_lists(original_list, updated_list, key):
+    """
+    Replace values in a list with those in an updated list.
+    Example:
+    >>> original = [{'id': '1', 'updated': 'n'}, {'id': '2', 'updated': 'n'}, {'id': '11', 'updated': 'n'}]
+    >>> updated = [{'id': '1', 'updated': 'y'}, {'id': '3', 'updated': 'y'}, {'id': '11', 'updated': 'n',
+    >>>                                                                                             'remove': True}]
+    >>> result = [{'id': '1', 'updated': 'y'}, {'id': '2', 'updated': 'n'}, {'id': '3', 'updated': 'y'}]
+
+    :type original_list: ``list``
+    :param original_list: The original list.
+
+    :type updated_list: ``list``
+    :param updated_list: The updated list.
+
+    :type key: ``str``
+    :param key: The key to replace elements by.
+
+    :rtype: ``list``
+    :return: The merged list.
+
+    """
+
+    original_dict = {element[key]: element for element in original_list}
+    updated_dict = {element[key]: element for element in updated_list}
+    original_dict.update(updated_dict)
+
+    removed = [obj for obj in original_dict.values() if obj.get('remove', False) is True]
+    for r in removed:
+        demisto.debug('Removing from integration context: {}'.format(str(r)))
+
+    merged_list = [obj for obj in original_dict.values() if obj.get('remove', False) is False]
+
+    return merged_list
+
+
+def set_integration_context(context, sync=True, version=-1):
+    """
+    Sets the integration context.
+
+    :type context: ``dict``
+    :param context: The context to set.
+
+    :type sync: ``bool``
+    :param sync: Whether to save the context directly to the DB.
+
+    :type version: ``int``
+    :param version: The version of the context to set.
+
+    :rtype: ``dict``
+    :return: The new integration context
+    """
+    demisto.debug('Setting integration context {}:'.format(str(context)))
+    if is_versioned_context_available():
+        context['version'] = str(version + 1)
+        demisto.debug('Updating integration context to version {}. Sync: {}'.format(version + 1, sync))
+        return demisto.setIntegrationContextVersioned(context, version, sync)
+    else:
+        return demisto.setIntegrationContext(context)
+
+
+def get_integration_context(sync=True, with_version=False):
+    """
+    Gets the integration context.
+
+    :type sync: ``bool``
+    :param sync: Whether to get the integration context directly from the DB.
+
+    :type with_version: ``bool``
+    :param with_version: Whether to return the version.
+
+    :rtype: ``dict``
+    :return: The integration context.
+    """
+    if is_versioned_context_available():
+        integration_context = demisto.getIntegrationContextVersioned(sync)
+
+        if with_version:
+            return integration_context
+        else:
+            return integration_context.get('context', {})
+    else:
+        return demisto.getIntegrationContext()
+
+
+def is_versioned_context_available():
+    """
+    Determines whether versioned integration context is available according to the server version.
+
+    :rtype: ``bool``
+    :return: Whether versioned integration context is available
+    """
+    return is_demisto_version_ge(MIN_VERSION_FOR_VERSIONED_CONTEXT)
+
+
+def set_to_integration_context_with_retries(context, object_keys=None, sync=True,
+                                            max_retry_times=CONTEXT_UPDATE_RETRY_TIMES):
+    """
+    Update the integration context with a dictionary of keys and values with multiple attempts.
+    The function supports merging the context keys using the provided object_keys parameter.
+    If the version is too old by the time the context is set,
+    another attempt will be made until the limit after a random sleep.
+
+    :type context: ``dict``
+    :param context: A dictionary of keys and values to set.
+
+    :type object_keys: ``dict``
+    :param object_keys: A dictionary to map between context keys and their unique ID for merging them.
+
+    :type sync: ``bool``
+    :param sync: Whether to save the context directly to the DB.
+
+    :type max_retry_times: ``int``
+    :param max_retry_times: The maximum number of attempts to try.
+
+    :rtype: ``None``
+    :return: None
+    """
+    attempt = 0
+
+    # do while...
+    while True:
+        if attempt == max_retry_times:
+            raise Exception('Failed updating integration context. Max retry attempts exceeded.')
+
+        # Update the latest context and get the new version
+        integration_context, version = update_integration_context(context, object_keys, sync)
+
+        demisto.debug('Attempting to update the integration context with version {}.'.format(version))
+
+        # Attempt to update integration context with a version.
+        # If we get a ValueError (DB Version), then the version was not updated and we need to try again.
+        attempt += 1
+        try:
+            set_integration_context(integration_context, sync, version)
+            demisto.debug('Successfully updated integration context. New version is {}.'
+                          ''.format(version + 1 if version != -1 else version))
+            break
+        except ValueError as ve:
+            demisto.debug('Failed updating integration context with version {}: {} Attempts left - {}'
+                          ''.format(version, str(ve), CONTEXT_UPDATE_RETRY_TIMES - attempt))
+            # Sleep for a random time
+            time_to_sleep = randint(1, 100) / 1000
+            time.sleep(time_to_sleep)
+
+
+def get_integration_context_with_version(sync=True):
+    """
+    Get the latest integration context with version, if available.
+
+    :type sync: ``bool``
+    :param sync: Whether to get the context directly from the DB.
+
+    :rtype: ``tuple``
+    :return: The latest integration context with version.
+    """
+    latest_integration_context_versioned = get_integration_context(sync, with_version=True)
+    version = -1
+    if is_versioned_context_available():
+        integration_context = latest_integration_context_versioned.get('context', {})
+        if sync:
+            version = latest_integration_context_versioned.get('version', 0)
+    else:
+        integration_context = latest_integration_context_versioned
+
+    return integration_context, version
+
+
+def update_integration_context(context, object_keys=None, sync=True):
+    """
+    Update the integration context with a given dictionary after merging it with the latest integration context.
+
+    :type context: ``dict``
+    :param context: The keys and values to update in the integration context.
+
+    :type object_keys: ``dict``
+    :param object_keys: A dictionary to map between context keys and their unique ID for merging them
+    with the latest context.
+
+    :type sync: ``bool``
+    :param sync: Whether to use the context directly from the DB.
+
+    :rtype: ``tuple``
+    :return: The updated integration context along with the current version.
+
+    """
+    integration_context, version = get_integration_context_with_version(sync)
+    if not object_keys:
+        object_keys = {}
+
+    for key, _ in context.items():
+        latest_object = json.loads(integration_context.get(key, '[]'))
+        updated_object = context[key]
+        if key in object_keys:
+            merged_list = merge_lists(latest_object, updated_object, object_keys[key])
+            integration_context[key] = json.dumps(merged_list)
+        else:
+            integration_context[key] = json.dumps(updated_object)
+
+    return integration_context, version
 
 
 class DemistoException(Exception):